+++ metabonding/src/rewards.rs
@@ -1,8 +1,11 @@
 elrond_wasm::imports!();
 elrond_wasm::derive_imports!();
 
-use crate::project::{Project, ProjectId};
-use core::ops::Deref;
+use crate::{
+    project::{Project, ProjectId},
+    validation::Signature,
+};
+use core::{borrow::Borrow, ops::Deref};
 
 pub type Week = usize;
 pub type PrettyRewards<M> =
@@ -36,7 +39,14 @@ impl<M: ManagedTypeApi> WeeklyRewards<M> {
 }
 
 #[elrond_wasm::module]
-pub trait RewardsModule: crate::project::ProjectModule + crate::util::UtilModule {
+pub trait RewardsModule:
+    elrond_wasm_modules::pause::PauseModule
+    + crate::project::ProjectModule
+    + crate::access_control::AccessControlModule
+    + crate::common_storage::CommonStorageModule
+    + crate::math::MathModule
+    + crate::validation::ValidationModule
+{
     #[endpoint(addRewardsCheckpoint)]
     fn add_rewards_checkpoint(
         &self,
@@ -69,10 +79,7 @@ pub trait RewardsModule: crate::project::ProjectModule + crate::util::UtilModule
         );
 
         let (payment_amount, payment_token) = self.call_value().payment_token_pair();
-        let project: Project<Self::Api> = match self.projects().get(&project_id) {
-            Some(p) => p,
-            None => sc_panic!("Invalid project ID"),
-        };
+        let project = self.get_project_or_panic(&project_id);
 
         let current_week = self.get_current_week();
         require!(
@@ -88,6 +95,54 @@ pub trait RewardsModule: crate::project::ProjectModule + crate::util::UtilModule
         require!(total_reward_supply == payment_amount, "Invalid amount");
 
         self.rewards_deposited(&project_id).set(&true);
+    }
+
+    #[endpoint(claimRewards)]
+    fn claim_rewards(
+        &self,
+        week: Week,
+        user_delegation_amount: BigUint,
+        user_lkmex_staked_amount: BigUint,
+        signature: Signature<Self::Api>,
+    ) {
+        require!(self.not_paused(), "May not claim rewards while paused");
+
+        let caller = self.blockchain().get_caller();
+        require!(
+            !self.rewards_claimed(&caller, week).get(),
+            "Already claimed rewards for this week"
+        );
+
+        let last_checkpoint_week = self.get_last_checkpoint_week();
+        require!(week <= last_checkpoint_week, "No checkpoint for week yet");
+
+        let checkpoint: RewardsCheckpoint<Self::Api> = self.rewards_checkpoints().get(week);
+        self.verify_signature(
+            week,
+            &caller,
+            &user_delegation_amount,
+            &user_lkmex_staked_amount,
+            &signature,
+        );
+
+        self.rewards_claimed(&caller, week).set(&true);
+
+        let weekly_rewards = self.get_rewards_for_week(
+            week,
+            &user_delegation_amount,
+            &user_lkmex_staked_amount,
+            &checkpoint.total_delegation_supply,
+            &checkpoint.total_lkmex_staked,
+        );
+        if !weekly_rewards.is_empty() {
+            for (id, payment) in weekly_rewards.iter() {
+                self.leftover_project_funds(id.borrow())
+                    .update(|leftover| *leftover -= &payment.amount);
+            }
+
+            self.send()
+                .direct_multi(&caller, &weekly_rewards.payments, &[]);
+        }
     }
 
     #[view(getUserClaimableWeeks)]
